% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClassesAccessors.R
\name{adjMatrix}
\alias{adjMatrix}
\alias{adjMatrix,prodeInput-method}
\title{adjMatrix}
\value{
and object of class Matrix.
}
\description{
This function allows to access the adjacency matrix in a \linkS4class{prodeInput}
object.
}
