% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClassesAccessors.R
\name{adjMatrixResults}
\alias{adjMatrixResults}
\alias{adjMatrixResults,prodeResults-method}
\title{adjMatrixResults}
\value{
and object of class Matrix.
}
\description{
This function allows to access the adjacency matrix in a \linkS4class{prodeResults}
object.
}
