% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClassesAccessors.R
\name{designMatrix}
\alias{designMatrix}
\alias{designMatrix,prodeInput-method}
\title{designMatrix}
\value{
and object of class Matrix.
}
\description{
This function allows to access the design matrix in a \linkS4class{prodeInput}
object.
}
