% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClassesAccessors.R
\name{filteredGenes}
\alias{filteredGenes}
\alias{filteredGenes,prodeResults-method}
\title{filteredGenes}
\value{
a DataFrame object.
}
\description{
This function allows to access the genes filtered out in a \linkS4class{prodeResults}
object.
}
