% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProdeInput.R
\name{getProdeInput}
\alias{getProdeInput}
\title{Constructs input object for PRODE run.}
\usage{
getProdeInput(score_matrix, col_data, edge_table, design = NULL)
}
\arguments{
\item{score_matrix}{matrix of input scores with genes on rows
and samples on columns. No \code{NA} values are allowed. Gene names are required
as \code{rownames}. Sample names are required as \code{colnames} and need to be
identical to \code{col_data} column names.}

\item{col_data}{data.frame of column data, as described by \linkS4class{SummarizedExperiment}. Row names
need to be identical to score-matrix column names and correspond to samples.
Every sample-level included in design formula (if provided) should be included in \code{col_data}.
The group variable, encoding the sample groups that need to be compared for NICE score
computation, has to be a binary integer vector, encoding 1 for the case-group and 0 for the
control-group.}

\item{edge_table}{matrix of interactions between gene pairs. This matrix displays
two columns for each gene in the pair. Genes not present in \code{score_matrix} row names
will be discarded.}

\item{design}{this is the design formula required for NICE score computation. In the case of
NIE score computation, this can be ignored (default set to \code{NULL}). \code{design} has to
be a \code{formula} object. Last variable in formula will be considered as the group
variable for the linear model fit, requiring 1 for case group and 0 for control group.}
}
\value{
An object of class \linkS4class{prodeInput}.
}
\description{
Constructs input object for PRODE run.
}
\details{
This function computes an adjacency matrix from edge list and filters
out all genes that are not present in the edge-list or score matrix. The resulting
\linkS4class{prodeInput} object inherits from \linkS4class{SummarizedExperiment}.
}
