% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rra_stats_functions.R
\name{getRandomBetas}
\alias{getRandomBetas}
\title{Runs linear models on input data}
\usage{
getRandomBetas(adj_m, n_iter)
}
\arguments{
\item{nn}{}
}
\value{
matrix with summary statistics of each fit
}
\description{
Runs linear models on input data
}
