% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rra_stats_functions.R
\name{getRandomBetasPar}
\alias{getRandomBetasPar}
\title{Runs linear models on input data}
\usage{
getRandomBetasPar(degs, n_iter, nn, cores = 1)
}
\arguments{
\item{covs}{}
}
\value{
matrix with summary statistics of each fit
}
\description{
Runs linear models on input data
}
