% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClassesAccessors.R
\name{modality}
\alias{modality}
\alias{modality,prodeInput-method}
\title{modality}
\value{
a character string.
}
\description{
This function allows to access modality in a \linkS4class{prodeInput}
object.
}
