% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{prodeInput-class}
\alias{prodeInput-class}
\alias{.prodeInput}
\title{\code{prodeInput} class}
\description{
prodeInput class extends \linkS4class{SummarizedExperiment} class
including adjacency matrix computed by \code{getProdeInput} function from
input \code{edge_list} and \code{modality}, a character string reporting
the type of score that needs to be computed by PRODE (\code{NIE_score} or
\code{NICE_score}).
}
