% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{prodeResults-class}
\alias{prodeResults-class}
\alias{.prodeResults}
\title{\code{prodeResults} class}
\description{
\code{prodeResults} class is an object which extends DataFrame
object (reporting results of PRODE run), by including \code{adjMatrix} as the
adjacency matrix resulting after PRODE run (in case some genes have been filtered out) and
filterdData, a DFrame object containing scores and statistics of genes that have been filtered
out during the analyisis (depending on the analyses settings).
}
