% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClassesAccessors.R
\name{results}
\alias{results}
\alias{results,prodeResults-method}
\title{results}
\value{
a  DataFrame object.
}
\description{
This function allows to access results in a \linkS4class{prodeResults}
object. Details on columns are reported in \code{\link{runProde}}
}
