% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClassesAccessors.R
\name{weights}
\alias{weights}
\alias{weights,prodeInput-method}
\alias{modality}
\title{weights}
\value{
a character string.
}
\description{
This function allows to access modality in a \linkS4class{prodeInput}
object.
}
