% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitLms.R
\name{fitLms}
\alias{fitLms}
\title{Fit linear models on each row of score-matrix}
\usage{
fitLms(x, y, extendedNICEStats = F)
}
\arguments{
\item{x}{a model.matrix object containing variables included in the linear model.
This object is expected to have rows corresponding to the score-matrix columns.}

\item{y}{a score-matrix with a number of columns equal to x, no NA values are allowed.}

\item{extendedNICEStats}{a logical if extended stats need to be computed.}
}
\value{
matrix object with statistics of linear model fits.
}
\description{
Fit linear models on each row of score-matrix
}
\examples{
\dontrun{
    y <- rnorm(100, 10, 10)
    x <- sample(c(1,0), 10, replace=T)
    fitLms(x, y)
}

}
