% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_models_fit_functions.R
\name{fitLms}
\alias{fitLms}
\title{Runs linear models on given as input a prode object.}
\usage{
fitLms(x, y)
}
\arguments{
\item{covs}{}
}
\value{
matrix with summary statistics of each fit
}
\description{
Runs linear models on given as input a prode object.
}
