% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rra_stats_functions.R
\name{getRandomBetas}
\alias{getRandomBetas}
\title{Runs linear models on input data}
\usage{
getRandomBetas(degs, n_iter, nn)
}
\arguments{
\item{covs}{}
}
\value{
matrix with summary statistics of each fit
}
\description{
Runs linear models on input data
}
