% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rra_stats_functions.R
\name{getRealBetas}
\alias{getRealBetas}
\title{Runs linear models on input data}
\usage{
getRealBetas(bet_tab, adj_m, back_dis)
}
\arguments{
\item{covs}{}
}
\value{
matrix with summary statistics of each fit
}
\description{
Runs linear models on input data
}
