% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRhos.R, R/rra_stats_functions.R
\name{getRealRhosFitDistr}
\alias{getRealRhosFitDistr}
\title{Computes \eqn{\rho} p-values according to RRA.}
\usage{
getRealRhosFitDistr(bet_tab, adj_m, back_par, runEss, scaledEst)

getRealRhosFitDistr(bet_tab, adj_m, back_par, runEss, scaledEst)
}
\arguments{
\item{bet_tab}{a table of linear models fit statistics as computed by \code{fitLms()}.}

\item{adj_m}{an adjacency matrix with rows and columns corresponding to rows of \code{beta_tab}.}

\item{back_par}{a table of parameters of background fitted distributions.}

\item{scaledEst}{logical, whether to used scaled coefficients or not to computee gene-level signal.}

\item{beta_tab}{a table of linear models fit statistics as computed by \code{fitLms()}.}
}
\value{
matrix object with statistics of RRA for each gen in beta_tab.

matrix object with statistics of RRA for each gen in beta_tab.
}
\description{
Computes \eqn{\rho} p-values according to RRA.

For each gene's neighborhood, this function computes the rho value
as described by the RRA algorithm (Kolde et al., 2012) and a p-value
leveraging an input list of fitted background rhos distributions on randomly
shuffled ranks, depending on neighborhood size. The lower the p-value,
the more genes in neighborhood are distributed towards low ranks.
}
\details{
For each gene's neighborhood, this function computes the rho value
as described by the RRA algorithm (Kolde et al., 2012) and a p-value
leveraging an input list of fitted background rhos distributions on randomly
shuffled ranks, depending on neighborhood size. The lower the p-value,
the more genes in neighborhood are distributed towards low ranks.
}
