% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_prode_run_function.R
\name{runProde}
\alias{runProde}
\title{Runs linear models on input data}
\usage{
runProde(
  scores_matrix,
  columns_data,
  condition,
  covariates,
  edge_table,
  cores = 1,
  filterCtrl = T
)
}
\arguments{
\item{covs}{}
}
\value{
matrix with summary statistics of each fit
}
\description{
Runs linear models on input data
}
