% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runProde.R
\name{runProde}
\alias{runProde}
\title{Run ProDe on pre-processed input data}
\usage{
runProde(
  prodeInput,
  cores = 1,
  filterCtrl = T,
  n_iter = 10000,
  extendedStats = F,
  computeBack = F
)
}
\arguments{
\item{prodeInput}{object of class ProdeInput. This is generated by a first
running \code{getProdeInput} function.}

\item{cores}{integer, the number of cores, default=1 (no parallelization)}

\item{filterCtrl}{logical, whether or not to filter genes depending on
mean of control samples, before running RRA algorithm.}

\item{n_iter}{integer, number of random iterations to generate the background
null distribution of random beta probabilities.}

\item{extendedStats}{logical, wether to computee extended per-group
stats on score matrix.}
}
\description{
Run ProDe on pre-processed input data
}
